@@ -2,43 +2,36 @@
   "name": "react-client",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://github.com/Georgy87/SPLICE_PL_CLIENT-REACT.git",
   "dependencies": {
-    "@ant-design/icons": "^4.7.0",
-    "@hookform/resolvers": "^2.8.8",
-    "@material-ui/icons": "^4.11.2",
-    "@reduxjs/toolkit": "^1.8.0",
-    "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
+    "@ant-design/icons": "^4.8.0",
+    "@hookform/resolvers": "^2.9.10",
+    "@reduxjs/toolkit": "^1.9.1",
     "@types/howler": "^2.2.7",
-    "@types/jest": "^27.4.1",
     "@types/material-ui": "^0.21.12",
-    "@types/node": "^17.0.23",
-    "@types/offscreencanvas": "^2019.6.4",
-    "@types/react": "^17.0.43",
-    "@types/react-dom": "^17.0.14",
-    "@types/react-redux": "^7.1.23",
+    "@types/node": "^17.0.45",
+    "@types/offscreencanvas": "^2019.7.0",
+    "@types/react": "^17.0.52",
+    "@types/react-dom": "^18.0.9",
     "@types/react-router-dom": "^5.3.3",
-    "antd": "^4.19.3",
+    "antd": "^4.24.5",
     "axios": "^0.26.1",
-    "axios-mock-adapter": "^1.20.0",
-    "classnames": "^2.3.1",
-    "easy-yandex-s3": "^1.1.6",
+    "axios-mock-adapter": "^1.21.2",
+    "classnames": "^2.3.2",
+    "easy-yandex-s3": "^1.1.8",
     "eslint-config-prettier": "^8.5.0",
     "howler": "^2.2.3",
-    "node-sass": "^7.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^7.28.1",
-    "react-icons": "^4.3.1",
-    "react-image-crop": "^9.0.5",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "^5.0.0",
-    "redux": "^4.1.2",
-    "redux-thunk": "^2.4.1",
-    "typescript": "^4.6.3",
+    "node-sass": "^7.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.40.0",
+    "react-icons": "^4.7.1",
+    "react-image-crop": "^9.1.1",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.4.5",
+    "react-scripts": "^5.0.1",
+    "redux": "^4.2.0",
+    "redux-thunk": "^2.4.2",
+    "typescript": "^4.9.4",
     "uuid": "^8.3.2",
     "web-vitals": "^2.1.4",
     "yup": "^0.32.11"
@@ -70,9 +63,20 @@
     ]
   },
   "devDependencies": {
+    "@testing-library/dom": "^8.19.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
     "@types/enzyme": "^3.10.12",
+    "@types/jest": "^28.1.8",
+    "@types/redux-mock-store": "^1.0.3",
     "@types/uuid": "^8.3.4",
     "gh-pages": "^4.0.0",
-    "redux-mock-store": "^1.5.4"
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
+    "jest-fetch-mock": "^3.0.3",
+    "redux-mock-store": "^1.5.4",
+    "ts-jest": "^28.0.8"
   }
 }
