@@ -1,23 +1,23 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 
-import reportWebVitals from './reportWebVitals';
 import { store } from './store';
 import { App } from './App';
+import reportWebVitals from './reportWebVitals';
 
 import './index.css';
 
-ReactDOM.render(
-	<React.StrictMode>
-		<Provider store={store}>
-			<BrowserRouter>
-				<App />
-			</BrowserRouter>
-		</Provider>
-	</React.StrictMode>,
-	document.getElementById('root'),
+const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
+root.render(
+    // <React.StrictMode>
+        <Provider store={store}>
+            <BrowserRouter>
+                <App />
+            </BrowserRouter>
+        </Provider>
+    // </React.StrictMode>
 );
 
 reportWebVitals();
