@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
 import * as Icons from 'react-icons/fa';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 
 import { NavbarList } from '../NavbarList/NavbarList';
 import { Sidebar } from '../Sidebar/Sidebar';
@@ -11,59 +11,60 @@ import { selectAuth, selectUser } from '../../store/selectors/userSelectors';
 import { logout } from '../../store/slices/user/userSlice';
 import { setDefaultPackState } from '../../store/slices/pack/packSlice';
 import defaultAvatar from '../../assets/avatar/unnamed.jpg';
+import { useAppDispatch } from '../../store/types';
 
 import styles from './Navbar.module.scss';
 
 export const Navbar = () => {
-	const [sidebar, setSideBar] = useState(false);
-	const user = useSelector(selectUser);
+    const [sidebar, setSideBar] = useState(false);
+    const user = useSelector(selectUser);
 
-	const dispatch = useDispatch();
+    const dispatch = useAppDispatch();
 
-	const isAuth = useSelector(selectAuth);
-	return (
-		<>
-			<nav className={isAuth ? styles.navbar : `${styles.navbar} ${styles.notAuth}`}>
-				<Link to='/' data-testid='packs-link' className={styles.navbarLogo} onClick={() => setSideBar(true)}>
-					<IconLayout iconName={'music'} />
-					SampleCloud
-				</Link>
+    const isAuth = useSelector(selectAuth);
+    return (
+        <>
+            <nav className={isAuth ? styles.navbar : `${styles.navbar} ${styles.notAuth}`}>
+                <Link to="/" data-testid="packs-link" className={styles.navbarLogo} onClick={() => setSideBar(true)}>
+                    <IconLayout iconName={'music'} />
+                    SampleCloud
+                </Link>
 
-				{sidebar ? (
-					<Icons.FaTimes className={styles.sidebarToggleLogo} onClick={() => setSideBar(!sidebar)} />
-				) : (
-					<Icons.FaBars className={styles.sidebarToggleLogo} onClick={() => setSideBar(!sidebar)} />
-				)}
+                {sidebar ? (
+                    <Icons.FaTimes className={styles.sidebarToggleLogo} onClick={() => setSideBar(!sidebar)} />
+                ) : (
+                    <Icons.FaBars className={styles.sidebarToggleLogo} onClick={() => setSideBar(!sidebar)} />
+                )}
 
-				{isAuth ? <NavbarList /> : null}
-				{isAuth && (
-					<div className={styles.avatar}>
-						<img src={user?.avatar ? user.avatar : defaultAvatar} alt='user-avatar' />
-					</div>
-				)}
-				{
-					<Link to='/login' data-testid='login-link'>
-						{isAuth ? (
-							<ButtonLayout
-								typeStyle={'sign-in-out'}
-								onClicked={() => {
-									dispatch(logout());
-									dispatch(setDefaultPackState());
-								}}
-							>
-								<IconLayout iconName={'login'} />
-								<span>Log Out</span>
-							</ButtonLayout>
-						) : (
-							<ButtonLayout typeStyle={'sign-in-out'}>
-								<IconLayout iconName={'logout'} />
-								<span>Log In</span>
-							</ButtonLayout>
-						)}
-					</Link>
-				}
-			</nav>
-			{sidebar && <Sidebar sidebar={sidebar} setSideBar={setSideBar} />}
-		</>
-	);
+                {isAuth ? <NavbarList /> : null}
+                {isAuth && (
+                    <div className={styles.avatar}>
+                        <img src={user?.avatar ? user.avatar : defaultAvatar} alt="user-avatar" />
+                    </div>
+                )}
+                {
+                    <Link to="/login" data-testid="login-link">
+                        {isAuth ? (
+                            <ButtonLayout
+                                typeStyle={'sign-in-out'}
+                                onClicked={() => {
+                                    dispatch(logout());
+                                    dispatch(setDefaultPackState());
+                                }}
+                            >
+                                <IconLayout iconName={'login'} />
+                                <span>Log Out</span>
+                            </ButtonLayout>
+                        ) : (
+                            <ButtonLayout typeStyle={'sign-in-out'}>
+                                <IconLayout iconName={'logout'} />
+                                <span>Log In</span>
+                            </ButtonLayout>
+                        )}
+                    </Link>
+                }
+            </nav>
+            {sidebar && <Sidebar sidebar={sidebar} setSideBar={setSideBar} />}
+        </>
+    );
 };
