@@ -1,13 +1,12 @@
 import React, { useEffect, useRef, useState } from 'react';
-import { useDispatch } from 'react-redux';
 
+import { useAppDispatch } from '../../store/types';
 import { useSound } from '../../hooks/useSound';
 import { IconChangeLayout } from '../../layouts/IconChangeLayout/IconChangeLayout';
 import { SampleSliderLayout } from '../../layouts/SampleSliderLayout/SampleSliderLayout';
 import { formatTime } from '../../utils/formatTime';
 import { Samples } from '../../store/slices/samples/types';
 import { IconLayout } from '../../layouts/IconLayout/IconLayout';
-
 import { canvasSampleService } from '../../services/canvasSampleService';
 import { AddSampleInfoModal } from '../AddSampleInfoModal/AddSampleInfoModal';
 import { fetchDeleteLike, fetchSetLike } from '../../store/slices/samples/actions';
@@ -28,7 +27,7 @@ export const SampleItem: React.FC<PropsType> = ({ sample, idx }) => {
 
 	const { playTrack, isPlaying, currentSampleId, currentTime, percent } = useSound();
 
-	const dispatch = useDispatch();
+	const dispatch = useAppDispatch();
 
 	const canvasRef = useRef<HTMLCanvasElement>(null);
 
