@@ -1,8 +1,8 @@
 import axios, { AxiosRequestConfig } from 'axios';
 
 const instance = axios.create({
-    // baseURL: 'http://localhost:8000/',
-    baseURL: 'https://sample-cloud-server.herokuapp.com/',
+    baseURL: 'http://localhost:8000/',
+    // baseURL: 'https://sample-cloud-server.herokuapp.com/',
     // baseURL: 'https://sample-platform-servergpq.onrender.com'
 });
 
