@@ -5,6 +5,16 @@ import { packsReducer } from './slices/pack/packSlice';
 import { samplesReducer } from './slices/samples/samplesSlice';
 import { userReducer } from './slices/user/userSlice';
 
+// export const store = configureStore({
+// 	reducer: {
+// 		packs: packSlice.reducer,
+// 		samples: samplesSlice.reducer,
+// 		user: userSlice.reducer,
+// 	},
+// 	middleware: [thunk],
+// });
+
+
 export const rootReducer = combineReducers({
 	packs: packsReducer,
 	samples: samplesReducer,
@@ -13,5 +23,5 @@ export const rootReducer = combineReducers({
 
 export const store = configureStore({
 	reducer: rootReducer,
-	middleware: [thunk],
+	// middleware: [thunk],
 });
